Added creation of Unit for tests.
