+++ api/uniter/export_test.go
@@ -5,7 +5,10 @@ package uniter
 import (
 	"fmt"
 
+	"github.com/juju/names"
+
 	"github.com/juju/juju/api/base/testing"
+	"github.com/juju/juju/apiserver/params"
 )
 
 var (
@@ -34,4 +37,9 @@ func PatchUnitResponse(p testing.Patcher, u *Unit, expectedRequest string, respo
 // PatchUnitFacadeCall changes the internal FacadeCaller to one that calls the provided request handler function.
 func PatchUnitFacadeCall(p testing.Patcher, u *Unit, respFunc func(request string, params, response interface{}) error) {
 	testing.PatchFacadeCall(p, &u.st.facade, respFunc)
+}
+
+// CreateUnit creates uniter.Unit for tests.
+func CreateUnit(st *State, tag names.UnitTag) *Unit {
+	return &Unit{st, tag, params.Alive}
 }
